Module rnd::pseudo_random

This module provides a few different implementations of cryptographically insecure random number generators suitable for numeric simulations.

The default algorithm which is used in the thread-local generator ThreadLocalPrng is Tyge Løvset’s stc64 generator which is implemented in Stc64.

Another fast high quality algorithm is Blackman and Vigna’s (2019) xoshiro256** which is provided by XoShiRo256StarStar.

For applications that use tuples of consecutively generated values, it may be desirable to use a generator that is k-dimensionally equidistributed such that k is at least as large as the length of the tuples being generated. The generator Lcg64Xor1024Mix, which is a Rust port of Java’s L64X1024MixRandom algorithm is provably 16-dimensionally equidistributed. This generator has a much larger period (264(21024−1)) and state space (1088 bits) than the other generators and is about 3 to 4 times slower than Stc64.

All of these algorithms have good performance in statistical tests and so far no major issues are known. None of them is cryptographically secure. A weakness of the current implementation is that all of them can only be seeded by a single i64 which is theoretically insufficient for the state space these generators have. However, this should hardly be detectable in actual simulations.

Structs

Lcg64Xor1024Mix

The L64X1024MixRandom algorithm from JDK 17 which uses a linear congruential generator (LCG) as a first subgenerator and a Xor-based generator (xoroshiro1024) as a second subgenerator and then applies a 64-bit mixing function identified by Doug Lea.

Stc64

The 256-bit generator Stc64 is Tyge Løvset’s improved variation of Sfc64. See https://github.com/tylov/STC/blob/master/include/stc/crandom.h.

ThreadLocalPrng

A thread-local generator that wraps a Stc64 generator.

XoShiRo256StarStar

256-bit xoshiro256** pseudo random generator suggested by David Blackman and Sebastiano Vigna (2019). It is about 40% faster than XorShift64Star despite having a 4 times larger state space.

Traits

PseudoRandom

A generator of uniform pseudorandom values.