Trait rnd::pseudo_random::PseudoRandom

pub trait PseudoRandom {

    fn next_long(&mut self) -> i64;

    fn next_int(&mut self) -> i32 { ... }

    fn next_double(&mut self) -> f64 { ... }

    fn next_float(&mut self) -> f32 { ... }

    fn next_bool(&mut self) -> bool { ... }

    fn next_long_up_to(&mut self, n: i64) -> i64 { ... }

    fn next_int_up_to(&mut self, n: i32) -> i32 { ... }

    fn next_long_from_interval(&mut self, min: i64, max: i64) -> i64 { ... }

    fn next_int_from_interval(&mut self, min: i32, max: i32) -> i32 { ... }

    fn next_double_from_interval(&mut self, min: f64, max: f64) -> f64 { ... }

    fn next_float_from_interval(&mut self, min: f32, max: f32) -> f32 { ... }

    fn next_bytes(&mut self, bytes: &mut [u8]) { ... }

    fn next_longs(&mut self, longs: &mut [i64]) { ... }

    fn next_doubles(&mut self, doubles: &mut [f64]) { ... }

    fn next_gaussians(&mut self, out: &mut [f64; 2]) { ... }

}

A generator of uniform pseudorandom values.

Implementors need to supply an implementation of the next_long method.

The default implementations in this trait are efficient for methods that need at least 33 bits of randomness but somehow wasteful for the other methods because it dissipates valuable random bits piled up in the call to next_long() whenever less than 33 random bits are needed for the result type.

Required methods

fn next_long(&mut self) -> i64

Returns a uniformly distributed signed 64-bit integer.

Provided methods

fn next_int(&mut self) -> i32

Returns a uniformly distributed signed 32-bit integer.

fn next_double(&mut self) -> f64

Returns a uniformly distributed 64-bit floating point value.

fn next_float(&mut self) -> f32

Returns a uniformly distributed 32-bit floating point value.

fn next_bool(&mut self) -> bool

Returns an equi-distributed boolean value.

fn next_long_up_to(&mut self, n: i64) -> i64

Returns a uniformly distributed i64 value between 0 (inclusive) and n (exclusive) where n is the strictly positive bound on the random number to be returned. This method panics when n is <= 0.

Panics

Panics if n is zero or negative.

fn next_int_up_to(&mut self, n: i32) -> i32

Returns a uniformly distributed i32 value between 0 (inclusive) and n (exclusive) where n is the strictly positive bound on the random number to be returned. This method panics when n is <= 0.

Panics

Panics if n is zero or negative.

fn next_long_from_interval(&mut self, min: i64, max: i64) -> i64

Returns a signed 64-bit integer which is uniformly distributed in the interval [min, max].

Panics

Panics if max < min.

fn next_int_from_interval(&mut self, min: i32, max: i32) -> i32

Returns a signed 32-bit integer which is uniformly distributed in the interval [min, max].

Panics

Panics if max < min.

fn next_double_from_interval(&mut self, min: f64, max: f64) -> f64

Returns a 64-bit floating point value which is uniformly distributed in the interval [min, max).

fn next_float_from_interval(&mut self, min: f32, max: f32) -> f32

Returns a 32-bit floating point value which is uniformly distributed in the interval [min, max).

fn next_bytes(&mut self, bytes: &mut [u8])

Generates random bytes and places them into the user-supplied bytes slice. The number of random bytes produced is equal to the length of the slice.

fn next_longs(&mut self, longs: &mut [i64])

Generates random signed 64-bit integers and places them into the user-supplied longs slice. The number of random integers produced is equal to the length of the slice.

fn next_doubles(&mut self, doubles: &mut [f64])

Generates random 64-bit floating point values and places them into the user-supplied doubles slice. The number of random integers produced is equal to the length of the slice.

fn next_gaussians(&mut self, out: &mut [f64; 2])

Generates two standard normal distributed 64-bit floating point values and stores them into the user-supplied out array. Here, “standard normal” means N(0, 1), i.e. a normal distribution with expectation 0 and variance 1.